third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-keystatuses-multiple-keys.html

<!DOCTYPE html>
<html>
    <head>
        <title>Verify MediaKeySession.keyStatuses handles multiple keys</title>
        <script src="encrypted-media-utils.js"></script>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <script>
            async_test(function(test)
            {
                var mediaKeySession;
                var initDataType;
                var initData;

                // Create a list of multiple unique keyIds.
                var keyIds = [ stringToUint8Array('keyid'),
                               stringToUint8Array('a really long keyid'),
                               new Uint8Array([0,1,2,3]),
                               new Uint8Array([0,1,2,3,4]),
                               new Uint8Array([0,1,2,3,4,5]),
                               new Uint8Array([0,1,2,3,4,6]),
                               stringToUint8Array('7'),
                               stringToUint8Array('8'),
                               stringToUint8Array('9'),
                               stringToUint8Array('10'),
                               stringToUint8Array('11'),
                               stringToUint8Array('12'),
                               stringToUint8Array('13'),
                               stringToUint8Array('14'),
                               stringToUint8Array('15'),
                               stringToUint8Array('16'),
                               stringToUint8Array('17'),
                               stringToUint8Array('18'),
                               stringToUint8Array('19'),
                               stringToUint8Array('20'),
                               stringToUint8Array('21'),
                               stringToUint8Array('22'),
                               stringToUint8Array('23'),
                               stringToUint8Array('24'),
                               stringToUint8Array('25')
                             ];


                function processKeyStatusesChange(event)
                {
                    // All 25 keyIds should be returned in keyStatuses.
                    verifyKeyStatuses(mediaKeySession.keyStatuses, { expected: keyIds, unexpected: [] });
                    test.done();
                }

                function processMessage(event)
                {
                    // No keys added yet.
                    assert_equals(mediaKeySession.keyStatuses.size, 0);

                    waitForEventAndRunStep('keystatuseschange', mediaKeySession, processKeyStatusesChange, test);

                    // Use the same key for all 25 keyIds.
                    var rawKey = new Uint8Array([0xeb, 0xdd, 0x62, 0xf1, 0x68, 0x14, 0xd2, 0x7b,
                                                 0x68, 0xef, 0x12, 0x2a, 0xfc, 0xe4, 0xae, 0x3c]);
                    var jwks = keyIds.map(function(keyId) { return createJWK(keyId, rawKey); });
                    var jwkSet = stringToUint8Array(createJWKSet.apply(this, jwks));

                    mediaKeySession.update(jwkSet).catch(function(error) {
                        forceTestFailureFromPromise(test, error);
                    });
                }

                navigator.requestMediaKeySystemAccess('org.w3.clearkey', getSimpleConfiguration()).then(function(access) {
                    initDataType = access.getConfiguration().initDataTypes[0];
                    initData = getInitData(initDataType);
                    return access.createMediaKeys();
                }).then(function(mediaKeys) {
                    mediaKeySession = mediaKeys.createSession();

                    // There should be no keys defined yet.
                    assert_equals(mediaKeySession.keyStatuses.size, 0);

                    waitForEventAndRunStep('message', mediaKeySession, processMessage, test);
                    return mediaKeySession.generateRequest(initDataType, initData);
                }).catch(function(error) {
                    forceTestFailureFromPromise(test, error);
                });
            }, 'Verify MediaKeySession.keyStatuses handles multiple keys.');
        </script>
    </body>
</html>
